Graduate Labs Components - Part 2: Replace Popover/Tooltip with Popover2/Tooltip2 (#1891)
* Delete popover/ and tooltip/ (except SVG variants); rename ___2/ to ___/

* Rename 'POPOVER2' => 'POPOVER'

* Rename 'popover2' => 'popover'

* Move PopoverInteractionKind into new popover.tsx

* Rename 'Popover2' => 'Popover'

* Fix a bunch more 'Popover2' => 'Popover' renames

* Rename 'tooltip2' => 'tooltip'

* Rename 'Tooltip2' => 'Tooltip'

* Resolve popover2/tooltip2 filenames

* Whoops, add popover and tooltip CSS files again

* Fix compile errors

* Fix dist:apps via some isDisabled=>disabled references

* Set tag = false in <Manager>

* Try to fix submenu issues

* Delete minimal class, use minimal prop instead

* Add a <span> wrapper around popover targets/overlays, fix some submenu styles

* Set minimal=true in the docs navigator, instead of providing .pt-minimal className

* Move <MenuItem> minimal prop below the spread

* Add popoverRef prop

* Fix left-aligned menus

* Clarifying comment about ALIGN_LEFT

* Fix lingering popover2 mention in popover.md

* Delete unused ReactDOM import in <MenuItem>

* Fix context menus

* Fix Tooltip rootElementTag test, add POPOVER_WRAPPER class

* Delete SVGTooltip

Doesn't seem necessary, and test fails innately
because React 16 seems to freak out if you use
rootElementTag="g" without an <svg> wrapper that
you add yourself.

* Fix tooltip tests

* Fix popover tests

* Fix MenuItem tests (and skip a couple)

* Fix Popover tests

* Delete SVGPopover

* Fix skipped coverage in karma.conf.js

* Skip converage for popoverMigrationUtils

* Delete tetherUtils

* Fix lint

* Add Popover2 deprecated props to rename script

* Run rename script (isDisabled => disabled, useModal => hasBackdrop)

* Delete Popover2 deprecated-prop warnings

* Fix build

* Remove redundant/incorrect pt-popover-wrapper styling

* Fix DateInput tests

* Inline classNames for clarity

we have a naming pattern like [componentName][PropName],
but popoverClassName is already the name of a prop, so
it's ambiguous in this case.

* Fix flickering Popover in DateInput

* Fix DateInput tests

* Delete tether dependency from package.json

* Update lockfile (to remove tether)

* Change 'yarn lerna run test --scope' to 'cd ___ && yarn test'

* Add 'packages/*/node_modules/' to save_cache dirs

* No globs

* Cache bust: v5 => v6

* Undo singlRun flag change

* Revert changes to Circle config file

* Fix DateRangeInput test

* Fix table tests

* Clewis self-review cleanups

* Fix build (was due to errant find+replace), fix comment

* Fix inline popover targets (CSS)
