+++ packages/core/package.json
@@ -32,13 +32,11 @@
   "dependencies": {
     "@blueprintjs/icons": "^1.0.0",
     "@types/dom4": "^1.5.20",
-    "@types/tether": "^1.1.27",
     "classnames": "^2.2",
     "dom4": "^1.8",
     "normalize.css": "^7.0.0",
     "popper.js": "^1.12.6",
     "react-popper": "~0.7.4",
-    "tether": "^1.4",
     "tslib": "^1.5.0"
   },
   "peerDependencies": {
