+++ packages/core/src/components/components.md
@@ -21,7 +21,6 @@
 @page non-ideal-state
 @page overlay
 @page popover
-@page popover2
 @page portal
 @page progress
 @page sliders
@@ -32,5 +31,4 @@
 @page text
 @page toast
 @page tooltip
-@page tooltip2
 @page tree
