+++ packages/core/src/components/index.ts
@@ -48,11 +48,8 @@ export * from "./non-ideal-state/nonIdealState";
 export * from "./overlay/overlay";
 export * from "./text/text";
 export * from "./popover/popover";
-export * from "./popover2/popover2";
-export * from "./popover/svgPopover";
 export * from "./portal/portal";
 export * from "./progress/progressBar";
-export * from "./tooltip/svgTooltip";
 export * from "./slider/rangeSlider";
 export * from "./slider/slider";
 export * from "./spinner/spinner";
@@ -64,6 +61,5 @@ export * from "./tag-input/tagInput";
 export * from "./toast/toast";
 export * from "./toast/toaster";
 export * from "./tooltip/tooltip";
-export * from "./tooltip2/tooltip2";
 export * from "./tree/tree";
 export * from "./tree/treeNode";
