+++ packages/core/src/components/menu/_submenu.scss
@@ -6,29 +6,36 @@
 @import "./common";
 
 .pt-submenu {
-  > .pt-popover-target {
+  > .pt-popover-wrapper {
     display: inherit;
 
-    > .pt-menu-item {
-      padding-right: $pt-icon-size-standard + $pt-grid-size;
+    > .pt-popover-target {
+      display: inherit;
 
-      &::after {
-        @include pt-icon();
-        position: absolute;
-        right: $half-grid-size;
-        content: $pt-icon-caret-right;
+      > .pt-menu-item {
+        padding-right: $pt-icon-size-standard + $pt-grid-size;
 
-        .pt-large & {
-          line-height: $pt-icon-size-large;
+        &::after {
+          @include pt-icon();
+          position: absolute;
+          right: $half-grid-size;
+          content: $pt-icon-caret-right;
+
+          // need several layers of selectors to make sure we're targeting only
+          // the top level of menu items in the current submenu.
+          // stylelint-disable-next-line
+          .pt-large & {
+            line-height: $pt-icon-size-large;
+          }
         }
       }
     }
-  }
 
-  > .pt-popover-open > .pt-menu-item {
-    // keep a trail of hovered items as submenus are opened
-    // stylelint-disable-next-line scss/at-extend-no-missing-placeholder
-    @extend .pt-menu-item:hover;
+    > .pt-popover-open > .pt-menu-item {
+      // keep a trail of hovered items as submenus are opened
+      // stylelint-disable-next-line scss/at-extend-no-missing-placeholder
+      @extend .pt-menu-item:hover;
+    }
   }
 
   // minor adjustments to popover position so top submenu item lines up with submenu trigger item
