+++ packages/core/src/components/popover/_popover.scss
@@ -107,6 +107,12 @@ $popover-width: $pt-grid-size * 35 !default;
   }
 }
 
+.pt-popover-wrapper {
+  display: inline-block;
+  position: relative;
+  vertical-align: top;
+}
+
 .pt-popover-target {
   display: inline-block;
   position: relative;
