+++ packages/core/src/components/popover/popoverMigrationUtils.ts
@@ -12,6 +12,7 @@ import { Position } from "../../common/position";
  * @param position the position to convert
  */
 export function positionToPlacement(position: Position | "auto"): PopperJS.Placement {
+    /* istanbul ignore next */
     switch (position) {
         case Position.TOP_LEFT:
             return "top-start";
@@ -44,6 +45,7 @@ export function positionToPlacement(position: Position | "auto"): PopperJS.Place
     }
 }
 
+/* istanbul ignore next */
 function assertNever(x: never): never {
     throw new Error("Unexpected position: " + x);
 }
