+++ packages/core/src/components/tooltip/tooltip.tsx
@@ -5,13 +5,12 @@
  */
 
 import * as classNames from "classnames";
+import PopperJS from "popper.js";
 import * as React from "react";
 
 import * as Classes from "../../common/classes";
 import { Position } from "../../common/position";
 import { IIntentProps, IProps } from "../../common/props";
-import { ITetherConstraint } from "../../common/tetherUtils";
-
 import { Popover, PopoverInteractionKind } from "../popover/popover";
 
 export interface ITooltipProps extends IProps, IIntentProps {
@@ -21,17 +20,16 @@ export interface ITooltipProps extends IProps, IIntentProps {
     content: JSX.Element | string;
 
     /**
-     * Constraints for the underlying Tether instance.
-     * See http://github.hubspot.com/tether/#constraints
-     * @deprecated since v1.12.0; use `tetherOptions` instead.
-     */
-    constraints?: ITetherConstraint[];
-
-    /**
-     * Whether the tooltip is initially open.
+     * Initial opened state when uncontrolled.
      * @default false
      */
     defaultIsOpen?: boolean;
+
+    /**
+     * Prevents the popover from appearing when `true`.
+     * @default false
+     */
+    disabled?: boolean;
 
     /**
      * The amount of time in milliseconds the tooltip should remain open after the
@@ -63,17 +61,17 @@ export interface ITooltipProps extends IProps, IIntentProps {
     inline?: boolean;
 
     /**
-     * Prevents the tooltip from appearing when `true`.
-     * @default false
-     */
-    isDisabled?: boolean;
-
-    /**
-     * Whether or not the tooltip is visible. Passing this property will put the tooltip in
+     * Whether the popover is visible. Passing this prop puts the popover in
      * controlled mode, where the only way to change visibility is by updating this property.
      * @default undefined
      */
     isOpen?: boolean;
+
+    /**
+     * Popper modifier options, passed directly to internal Popper instance.
+     * See https://popper.js.org/popper-documentation.html#modifiers for complete details.
+     */
+    modifiers?: PopperJS.Modifiers;
 
     /**
      * Callback invoked in controlled mode when the tooltip open state *would* change due to
@@ -95,25 +93,23 @@ export interface ITooltipProps extends IProps, IIntentProps {
     portalClassName?: string;
 
     /**
-     * The position (relative to the target) at which the tooltip should appear.
-     * @default Position.TOP
+     * The position (relative to the target) at which the popover should appear.
+     *
+     * The default value of `"auto"` will choose the best position when opened
+     * and will allow the popover to reposition itself to remain onscreen as the
+     * user scrolls around.
+     * @default "auto"
      */
-    position?: Position;
+    position?: Position | "auto";
 
     /**
-     * The name of the HTML tag to use when rendering the tooltip target wrapper element.
+     * The name of the HTML tag to use when rendering the popover target wrapper element (`.pt-popover-target`).
      * @default "span"
      */
     rootElementTag?: string;
 
     /**
-     * Options for the underlying Tether instance.
-     * See http://tether.io/#options
-     */
-    tetherOptions?: Partial<Tether.ITetherOptions>;
-
-    /**
-     * A space-delimited string of class names that are applied to the tooltip (but not the target).
+     * A space-delimited string of class names that are applied to the tooltip.
      */
     tooltipClassName?: string;
 
@@ -125,51 +121,32 @@ export interface ITooltipProps extends IProps, IIntentProps {
      * @default 100
      */
     transitionDuration?: number;
-
-    /**
-     * Whether the arrow's offset should be computed such that it always points at the center
-     * of the target. If `false`, arrow position is hardcoded via CSS, which expects a 30px target.
-     * @default true
-     */
-    useSmartArrowPositioning?: boolean;
-
-    /**
-     * Whether the tooltip will try to reposition itself
-     * if there isn't room for it in its current position.
-     * @default false
-     */
-    useSmartPositioning?: boolean;
 }
 
 export class Tooltip extends React.PureComponent<ITooltipProps, {}> {
-    public static defaultProps: Partial<ITooltipProps> = {
-        hoverCloseDelay: 0,
-        hoverOpenDelay: 100,
-        isDisabled: false,
-        openOnTargetFocus: true,
-        position: Position.TOP,
-        rootElementTag: "span",
-        transitionDuration: 100,
-        useSmartArrowPositioning: true,
-        useSmartPositioning: false,
-    };
-
     public static displayName = "Blueprint.Tooltip";
 
-    public render(): JSX.Element {
-        const { children, intent, openOnTargetFocus, tooltipClassName } = this.props;
+    public static defaultProps: Partial<ITooltipProps> = {
+        defaultIsOpen: false,
+        disabled: false,
+        hoverCloseDelay: 0,
+        hoverOpenDelay: 100,
+        openOnTargetFocus: true,
+        transitionDuration: 100,
+    };
+
+    public render() {
+        const { children, intent, tooltipClassName, ...restProps } = this.props;
         const classes = classNames(Classes.TOOLTIP, Classes.intentClass(intent), tooltipClassName);
 
         return (
             <Popover
-                {...this.props}
-                arrowSize={22}
+                {...restProps}
                 autoFocus={false}
                 canEscapeKeyClose={false}
                 enforceFocus={false}
                 interactionKind={PopoverInteractionKind.HOVER_TARGET_ONLY}
                 lazy={true}
-                openOnTargetFocus={openOnTargetFocus}
                 popoverClassName={classes}
             >
                 {children}
