--- packages/core/src/components/tooltip2/tooltip2.md
@@ -1,7 +1,0 @@
-@# Tooltip2
-
-`Tooltip2` simply passes its props to [`Popover2`](#labs/popover2) with some modified defaults. The public API is the same as [`Tooltip`](#core/components/tooltip), except for the changes noted in the `Popover2` documentation.
-
-@reactExample Tooltip2Example
-
-@interface ITooltip2Props
