+++ packages/core/test/index.ts
@@ -18,7 +18,6 @@ import "./card/cardTests";
 import "./collapse/collapseTests";
 import "./collapsible-list/collapsibleListTests";
 import "./common/propsTests.ts";
-import "./common/tetherUtilsTests.ts";
 import "./common/utils/compareUtilsTests.ts";
 import "./common/utilsTests.ts";
 import "./context-menu/contextMenuTests";
@@ -35,9 +34,7 @@ import "./icon/iconTests";
 import "./menu/menuTests";
 import "./non-ideal-state/nonIdealStateTests";
 import "./overlay/overlayTests";
-import "./popover/arrowsTests";
-// import "./popover/popoverTests";
-import "./popover2/popover2Tests";
+import "./popover/popoverTests";
 import "./portal/portalTests";
 import "./progress/progressBarTests";
 import "./slider/rangeSliderTests";
