+++ packages/docs-app/src/examples/core-examples/index.ts
@@ -26,7 +26,6 @@ export * from "./numericInputExtendedExample";
 export * from "./nonIdealStateExample";
 export * from "./overlayExample";
 export * from "./popoverExample";
-export * from "./popover2Example";
 export * from "./progressExample";
 export * from "./rangeSliderExample";
 export * from "./radioExample";
@@ -36,7 +35,6 @@ export * from "./tagInputExample";
 export * from "./textExample";
 export * from "./spinnerExample";
 export * from "./tabsExample";
-export * from "./tooltip2Example";
 export * from "./inputGroupExample";
 export * from "./tagExample";
 export * from "./toastExample";
