+++ packages/docs-app/src/examples/core-examples/popoverExample.tsx
@@ -1,21 +1,23 @@
 /*
- * Copyright 2015 Palantir Technologies, Inc. All rights reserved.
+ * Copyright 2017 Palantir Technologies, Inc. All rights reserved.
  *
  * Licensed under the terms of the LICENSE file distributed with this project.
  */
 
 import * as classNames from "classnames";
+import PopperJS from "popper.js";
 import * as React from "react";
 
 import {
+    Button,
     Classes,
-    ITetherConstraint,
+    Icon,
+    Intent,
     Menu,
     MenuDivider,
     MenuItem,
     Popover,
     PopoverInteractionKind,
-    Position,
     RadioGroup,
     Slider,
     Switch,
@@ -29,113 +31,113 @@ const INTERACTION_KINDS = [
     { label: "Hover (target only)", value: PopoverInteractionKind.HOVER_TARGET_ONLY.toString() },
 ];
 
-const CONSTRAINTS = [
-    {
-        label: "None",
-        value: "[]",
-    },
-    {
-        label: "Smart positioning",
-        value: JSON.stringify([
-            {
-                attachment: "together",
-                to: "scrollParent",
-            },
-        ]),
-    },
-    {
-        label: "Pin to window",
-        value: JSON.stringify([
-            {
-                attachment: "together",
-                pin: true,
-                to: "window",
-            },
-        ]),
-    },
-];
+const PLACEMENTS = PopperJS.placements.map((p: PopperJS.Placement) => (
+    <option key={p} value={p}>
+        {p}
+    </option>
+));
+
+const POPPER_DOCS = "https://popper.js.org/popper-documentation.html#modifiers";
 
 export interface IPopoverExampleState {
     canEscapeKeyClose?: boolean;
     exampleIndex?: number;
+    hasBackdrop?: boolean;
     inheritDarkTheme?: boolean;
     inline?: boolean;
     interactionKind?: PopoverInteractionKind;
-    isModal?: boolean;
-    position?: Position;
+    minimal?: boolean;
+    modifiers?: PopperJS.Modifiers;
+    placement?: PopperJS.Placement;
     sliderValue?: number;
-    tetherConstraints?: ITetherConstraint[];
-    useSmartArrowPositioning?: boolean;
 }
 
 export class PopoverExample extends BaseExample<IPopoverExampleState> {
     public state: IPopoverExampleState = {
         canEscapeKeyClose: true,
         exampleIndex: 0,
+        hasBackdrop: false,
         inheritDarkTheme: true,
         inline: false,
         interactionKind: PopoverInteractionKind.CLICK,
-        isModal: false,
-        position: Position.RIGHT,
+        minimal: false,
+        modifiers: {
+            arrow: { enabled: true },
+            flip: { enabled: true },
+            keepTogether: { enabled: true },
+            preventOverflow: { enabled: true, boundariesElement: "scrollParent" },
+        },
+        placement: "auto",
         sliderValue: 5,
-        tetherConstraints: [],
-        useSmartArrowPositioning: true,
     };
 
     protected className = "docs-popover-example";
 
-    private handleConstraintChange = handleStringChange(constraints => {
-        this.setState({ tetherConstraints: JSON.parse(constraints) });
-    });
     private handleExampleIndexChange = handleNumberChange(exampleIndex => this.setState({ exampleIndex }));
     private handleInteractionChange = handleNumberChange(interactionKind => {
-        const isModal = this.state.isModal && interactionKind === PopoverInteractionKind.CLICK;
-        this.setState({ interactionKind, isModal });
+        const hasBackdrop = this.state.hasBackdrop && interactionKind === PopoverInteractionKind.CLICK;
+        this.setState({ interactionKind, hasBackdrop });
     });
-    private handlePositionChange = handleNumberChange(position => this.setState({ position }));
 
-    private toggleArrows = handleBooleanChange(useSmartArrowPositioning => {
-        this.setState({ useSmartArrowPositioning });
-    });
+    private handlePlacementChange = handleStringChange((placement: PopperJS.Placement) => this.setState({ placement }));
+    private handleBoundaryChange = handleStringChange((boundary: PopperJS.Boundary) =>
+        this.setState({
+            modifiers: {
+                ...this.state.modifiers,
+                preventOverflow: {
+                    boundariesElement: boundary,
+                    enabled: boundary.length > 0,
+                },
+            },
+        }),
+    );
+
     private toggleEscapeKey = handleBooleanChange(canEscapeKeyClose => this.setState({ canEscapeKeyClose }));
-    private toggleInheritDarkTheme = handleBooleanChange(inheritDarkTheme => this.setState({ inheritDarkTheme }));
     private toggleInline = handleBooleanChange(inline => {
         if (inline) {
             this.setState({
+                hasBackdrop: false,
                 inheritDarkTheme: false,
                 inline,
-                isModal: false,
-                tetherConstraints: [],
             });
         } else {
             this.setState({ inline });
         }
     });
-    private toggleModal = handleBooleanChange(isModal => this.setState({ isModal }));
+    private toggleMinimal = handleBooleanChange(minimal => this.setState({ minimal }));
 
     protected renderExample() {
-        const constraints = this.state.tetherConstraints;
-        const popoverClassName = classNames({
-            "pt-popover-content-sizing": this.state.exampleIndex <= 2,
+        const { exampleIndex, sliderValue, ...popoverProps } = this.state;
+        const popoverClassName = classNames(this.className, {
+            "pt-popover-content-sizing": exampleIndex <= 2,
         });
         return (
-            <Popover
-                content={this.getContents(this.state.exampleIndex)}
-                popoverClassName={popoverClassName}
-                tetherOptions={{ constraints }}
-                {...this.state}
-            >
-                <button className="pt-button pt-intent-primary">Popover target</button>
-            </Popover>
+            <div className="docs-popover-example-scroll" ref={this.centerScroll}>
+                <Popover popoverClassName={popoverClassName} portalClassName="foo" {...popoverProps}>
+                    <Button intent={Intent.PRIMARY} text="Popover target" />
+                    {this.getContents(exampleIndex)}
+                </Popover>
+                <p>
+                    Scroll around this container to experiment<br />
+                    with <code>flip</code> and <code>preventOverflow</code> modifiers.
+                </p>
+            </div>
         );
     }
 
     protected renderOptions() {
-        const isModalDisabled = this.state.inline || this.state.interactionKind !== PopoverInteractionKind.CLICK;
-        const inheritDarkThemeDisabled = this.state.inline;
-
+        const { arrow, flip, preventOverflow } = this.state.modifiers;
         return [
             [
+                <h5 key="app">Appearance</h5>,
+                <label className={Classes.LABEL} key="placement">
+                    Popover placement
+                    <div className={Classes.SELECT}>
+                        <select value={this.state.placement} onChange={this.handlePlacementChange}>
+                            {PLACEMENTS}
+                        </select>
+                    </div>
+                </label>,
                 <label className={Classes.LABEL} key="example">
                     Example content
                     <div className={Classes.SELECT}>
@@ -149,14 +151,16 @@ export class PopoverExample extends BaseExample<IPopoverExampleState> {
                         </select>
                     </div>
                 </label>,
+                <Switch checked={this.state.inline} label="Inline" key="inline" onChange={this.toggleInline} />,
                 <Switch
-                    checked={this.state.isModal}
-                    disabled={isModalDisabled}
-                    key="modal"
-                    label="Modal (requires Click interaction kind)"
-                    onChange={this.toggleModal}
+                    checked={this.state.minimal}
+                    label="Minimal (no arrow, simple transition)"
+                    key="minimal"
+                    onChange={this.toggleMinimal}
                 />,
-                <br key="break" />,
+            ],
+            [
+                <h5 key="int">Interactions</h5>,
                 <RadioGroup
                     key="interaction"
                     label="Interaction kind"
@@ -164,56 +168,51 @@ export class PopoverExample extends BaseExample<IPopoverExampleState> {
                     options={INTERACTION_KINDS}
                     onChange={this.handleInteractionChange}
                 />,
-            ],
-            [
-                <label className={Classes.LABEL} key="position">
-                    Popover position
-                    <div className={Classes.SELECT}>
-                        <select value={this.state.position} onChange={this.handlePositionChange}>
-                            <option value={Position.TOP.toString()}>Top</option>
-                            <option value={Position.TOP_RIGHT.toString()}>Top right</option>
-                            <option value={Position.RIGHT_TOP.toString()}>Right top</option>
-                            <option value={Position.RIGHT.toString()}>Right</option>
-                            <option value={Position.RIGHT_BOTTOM.toString()}>Right bottom</option>
-                            <option value={Position.BOTTOM_RIGHT.toString()}>Bottom right</option>
-                            <option value={Position.BOTTOM.toString()}>Bottom</option>
-                            <option value={Position.BOTTOM_LEFT.toString()}>Bottom left</option>
-                            <option value={Position.LEFT_BOTTOM.toString()}>Left bottom</option>
-                            <option value={Position.LEFT.toString()}>Left</option>
-                            <option value={Position.LEFT_TOP.toString()}>Left top</option>
-                            <option value={Position.TOP_LEFT.toString()}>Top left</option>
-                        </select>
-                    </div>
-                </label>,
-                <Switch
-                    checked={this.state.useSmartArrowPositioning}
-                    label="Smart arrow positioning"
-                    key="smartarrow"
-                    onChange={this.toggleArrows}
-                />,
                 <Switch
                     checked={this.state.canEscapeKeyClose}
                     label="Can escape key close"
                     key="escape"
                     onChange={this.toggleEscapeKey}
                 />,
-                <Switch checked={this.state.inline} label="Inline" key="inline" onChange={this.toggleInline} />,
-                <Switch
-                    checked={this.state.inheritDarkTheme}
-                    disabled={inheritDarkThemeDisabled}
-                    label="Should inherit dark theme"
-                    key="shouldinheritdarktheme"
-                    onChange={this.toggleInheritDarkTheme}
-                />,
                 <br key="break" />,
-                <RadioGroup
-                    disabled={this.state.inline}
-                    key="constraints"
-                    label="Constraints"
-                    onChange={this.handleConstraintChange}
-                    options={CONSTRAINTS}
-                    selectedValue={JSON.stringify(this.state.tetherConstraints)}
+            ],
+            [
+                <h5 key="mod">Modifiers</h5>,
+                <Switch
+                    checked={arrow.enabled}
+                    label="Arrow"
+                    key="arrow"
+                    onChange={this.getModifierChangeHandler("arrow")}
                 />,
+                <Switch
+                    checked={flip.enabled}
+                    label="Flip"
+                    key="flip"
+                    onChange={this.getModifierChangeHandler("flip")}
+                />,
+                <Switch
+                    checked={preventOverflow.enabled}
+                    label="Prevent overflow"
+                    key="preventOverflow"
+                    onChange={this.getModifierChangeHandler("preventOverflow")}
+                >
+                    <div className={Classes.SELECT} style={{ marginTop: 5 }}>
+                        <select
+                            disabled={!preventOverflow.enabled}
+                            value={preventOverflow.boundariesElement.toString()}
+                            onChange={this.handleBoundaryChange}
+                        >
+                            <option value="scrollParent">scrollParent</option>
+                            <option value="viewport">viewport</option>
+                            <option value="window">window</option>
+                        </select>
+                    </div>
+                </Switch>,
+                <p key="docs-link">
+                    <a href={POPPER_DOCS} target="_blank">
+                        Popper.js docs <Icon iconName="share" />
+                    </a>
+                </p>,
             ],
         ];
     }
@@ -258,4 +257,23 @@ export class PopoverExample extends BaseExample<IPopoverExampleState> {
     }
 
     private handleSliderChange = (value: number) => this.setState({ sliderValue: value });
+
+    private getModifierChangeHandler(name: keyof PopperJS.Modifiers) {
+        return handleBooleanChange(enabled => {
+            this.setState({
+                modifiers: {
+                    ...this.state.modifiers,
+                    [name]: { ...this.state.modifiers[name], enabled },
+                },
+            });
+        });
+    }
+
+    private centerScroll = (div: HTMLDivElement) => {
+        if (div != null) {
+            const container = div.parentElement;
+            container.scrollTop = div.clientHeight / 4;
+            container.scrollLeft = div.clientWidth / 4;
+        }
+    };
 }
