+++ packages/table/src/headers/columnHeaderCell.tsx
@@ -205,13 +205,6 @@ export class ColumnHeaderCell extends AbstractPureComponent<IColumnHeaderCellPro
             return undefined;
         }
 
-        const constraints = [
-            {
-                attachment: "together",
-                pin: true,
-                to: "window",
-            },
-        ];
         const classes = classNames(Classes.TABLE_TH_MENU_CONTAINER, {
             [Classes.TABLE_TH_MENU_OPEN]: this.state.isActive,
         });
@@ -223,13 +216,12 @@ export class ColumnHeaderCell extends AbstractPureComponent<IColumnHeaderCellPro
             <div className={classes}>
                 <div className={Classes.TABLE_TH_MENU_CONTAINER_BACKGROUND} />
                 <Popover
-                    tetherOptions={{ constraints }}
                     content={content}
                     position={Position.BOTTOM}
                     className={Classes.TABLE_TH_MENU}
+                    modifiers={{ preventOverflow: { boundariesElement: "window" } }}
                     popoverDidOpen={this.handlePopoverDidOpen}
                     popoverWillClose={this.handlePopoverWillClose}
-                    useSmartArrowPositioning={true}
                 >
                     <Icon iconName={menuIconName} />
                 </Popover>
