+++ packages/timezone/src/components/timezone-picker/timezonePicker.tsx
@@ -14,7 +14,7 @@ import {
     HTMLInputProps,
     IButtonProps,
     IInputGroupProps,
-    IPopover2Props,
+    IPopoverProps,
     IProps,
     MenuItem,
     Utils,
@@ -88,8 +88,8 @@ export interface ITimezonePickerProps extends IProps {
      */
     inputProps?: IInputGroupProps & HTMLInputProps;
 
-    /** Props to spread to `Popover2`. Note that `content` cannot be changed. */
-    popoverProps?: Partial<IPopover2Props> & object;
+    /** Props to spread to `Popover`. Note that `content` cannot be changed. */
+    popoverProps?: Partial<IPopoverProps> & object;
 }
 
 export interface ITimezonePickerState {
@@ -133,7 +133,7 @@ export class TimezonePicker extends AbstractPureComponent<ITimezonePickerProps, 
             placeholder: "Search for timezones...",
             ...inputProps,
         };
-        const finalPopoverProps: Partial<IPopover2Props> & object = {
+        const finalPopoverProps: Partial<IPopoverProps> & object = {
             ...popoverProps,
             popoverClassName: classNames(Classes.TIMEZONE_PICKER_POPOVER, popoverProps.popoverClassName),
         };
